+++ Sudoku.Benchmark/App.config
@@ -1,0 +1,15 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <runtime>
+    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
+      <dependentAssembly>
+        <assemblyIdentity name="System.Data.DataSetExtensions"
+                          publicKeyToken="b77a5c561934e089"
+                          culture="neutral" />
+        <!-- Assembly versions can be redirected in app,
+          publisher policy, or machine configuration files. -->
+        <bindingRedirect oldVersion="1.0.0.0-4.5.0.0" newVersion="4.6.26515.06" />
+      </dependentAssembly>
+    </assemblyBinding>
+  </runtime>
+</configuration>
