--- Sudoku.ChocoSolver/ChocoSolverJava/src/main/java/ChocoSolver.class
@@ -1,11 +1,0 @@
-public class ChocoSolverJava {
-    public static void PrintCells(int[][] cells) {
-        String cellsString = "";
-        for (int i = 0; i < cells.length; i++) {
-            for (int j = 0; j < cells[i].length; j++) {
-                cellsString += cells[i][j];
-            }
-        }
-        System.out.println(cellsString);
-    }
-}
